+++ src/test/java/com/startstepszalando/ecommerceshop/user/controller/UserControllerTest.java
@@ -1,7 +1,7 @@
 package com.startstepszalando.ecommerceshop.user.controller;
 
 import com.google.gson.Gson;
-import com.startstepszalando.ecommerceshop.exception.TokenRefreshException;
+import com.startstepszalando.ecommerceshop.exception.token.TokenRefreshException;
 import com.startstepszalando.ecommerceshop.exception.user.DuplicateUserException;
 import com.startstepszalando.ecommerceshop.jwt.JwtService;
 import com.startstepszalando.ecommerceshop.refreshToken.dto.TokenRefreshRequest;
